+++ src/main/java/de/maxhenkel/camera/ImageData.java
@@ -140,6 +140,7 @@ public class ImageData {
     }
 
     private static CompoundTag getImageTag(ItemStack stack) {
+        assert stack.getItem() instanceof ImageItem;
         CompoundTag compound = stack.getOrCreateTag();
 
         if (!compound.contains("image", Tag.TAG_COMPOUND)) {